Initial codebase (#1)
* Added project files

* Separated projects' logic

Now console interface's 'Main()' calls static 'Recognize(path)' function from image recognizer project. I also added a dummy check for exceptions in reading the image.

'var' keyword is now omitted in 'Recognize', which might be a source of extensive memory consumption (depending on how GC works). TODO: find a way to fix that.

* Changed logging a bit; still lacking a best practice async logging solution

* Small comment refactoring

* Changed project structure

* Basic directory traversing

* A small fix

* Replaced console in class lib with message queue

It's not parallel or async yet, but it's a step forward.

* Ignoring .onnx files

* Added parallel processing, but not printing

Output is still provided after the tasks have completed their job.

* Added exception handling for missing model

* Added asynchronous queue reading

It is still prone to errors and requires refactoring; that's the next major priority.

* Added sample images

Now that there are more images some problems are seen.

* Fixed result printing

Synchronized it with writing (using WritingPermission semaphor in console app), started printing the leftovers of the queue. Now the results are complete!

* Added proper cancellation

It needs additional testing, on a bigger dataset. I also added some new comments.

* Added readme-s

* Added more samples